requirements.txt
beautifulsoup4==4.12.2
pyTelegramBotAPI==4.9.0
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
Requests==2.30.0
schedule==1.1.0
